Заготовка Интерфейса администртора, связи комманд.

+++ db/migrate/20170307175112_create_xbet_bets.rb
@@ -6,7 +6,7 @@ class CreateXbetBets < ActiveRecord::Migration[5.0]
       t.float   :ratio
       t.boolean :active
 
-      t.integer :p
+      t.float   :p
       t.json    :pl
       #############
